update#2 of checkboxes
Update number 2 of checkboxes. Started working on the checkbox checked option which would disappear before. implemented the usecallback method. half of it is done. half remains as when clicked and the div that the checkboxes are inside of, are hidden and then shown, the check property is called to all checkboxes. will have to make some changes in the calling of {val.code}